# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lantern API
  description: |
    Canonical public API contract for the Lantern anonymous social client.

    All public content is authored by an anonymous persona. Persona objects never
    include email addresses, real-profile identifiers, or internal moderation notes.
    Real-profile data is available only through `/v1/me` to an authenticated session.
  version: "1.0.0"
  contact:
    name: Lantern Rebuild Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}:{port}"
    description: Local development server (default)
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://api-staging.lantern.app
    description: Staging server
  - url: https://api.lantern.app
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /v1/reports:
    post:
      operationId: createReport
      summary: Submit a report
      description: |
        Submits a report against a note, reply, or persona. The report is associated
        with the reporter's real profile for abuse prevention.
      tags:
        - Reports
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateRequest"
      responses:
        "201":
          description: Report submitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/reports/{reportId}:
    get:
      operationId: getReport
      summary: Get report status
      description: |
        Returns the status of a report submitted by the authenticated real profile.
        Moderator actions are not exposed beyond `resolved`.
      tags:
        - Reports
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/me/blocks:
    get:
      operationId: listBlocks
      summary: List blocked personas
      description: Returns the active persona's block list.
      tags:
        - Blocks
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated block list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageBlock"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createBlock
      summary: Block a persona
      description: |
        Blocks a persona directionally. The active persona will no longer see the
        target persona's content and the target cannot interact with the active
        persona's content.
      tags:
        - Blocks
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockCreateRequest"
      responses:
        "201":
          description: Block created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Block"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/me/blocks/{blockId}:
    delete:
      operationId: deleteBlock
      summary: Unblock a persona
      tags:
        - Blocks
      parameters:
        - name: blockId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Block removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/moderation/cases:
    get:
      operationId: listModerationCases
      summary: List moderation cases
      description: |
        Moderator-only endpoint. Lists open and recently resolved moderation cases.
      tags:
        - Moderation
      parameters:
        - name: status
          in: query
          description: Filter by case status.
          schema:
            type: string
            enum:
              - open
              - underReview
              - resolved
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated list of moderation cases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CursorPageModerationCase"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createModerationCase
      summary: Create a moderation case
      description: |
        Moderator-only endpoint. Creates a case from one or more reports.
      tags:
        - Moderation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationCaseCreateRequest"
      responses:
        "201":
          description: Case created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationCase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /v1/moderation/cases/{caseId}:
    get:
      operationId: getModerationCase
      summary: Get a moderation case
      description: Moderator-only endpoint.
      tags:
        - Moderation
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Moderation case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationCase"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateModerationCase
      summary: Update a moderation case
      description: |
        Moderator-only endpoint. Resolves the case by applying an outcome to the
        target content or account.
      tags:
        - Moderation
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationCaseUpdateRequest"
      responses:
        "200":
          description: Case updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationCase"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer access token returned by `POST /v1/auth/email-sessions` or `POST /v1/auth/refresh`.
  parameters:
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `pagination.nextCursor`.
      schema:
        type: string
        maxLength: 1024
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return. Default 20, maximum 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key for safely retrying mutating requests. The server
        stores the response for at least 24 hours and replays it for duplicate keys
        targeting the same endpoint and authenticated principal.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable description of the error.
        requestId:
          type: string
          format: uuid
          description: Request ID to include when contacting support.
...
    Avatar:
      type: object
      required:
        - seed
        - color
      properties:
        seed:
          type: string
          description: Deterministic seed used to render the generated avatar.
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          description: Hex color for the avatar background.
    Persona:
      type: object
      required:
        - id
        - alias
        - bio
        - avatar
        - createdAt
        - noteCount
      properties:
        id:
          type: string
          format: uuid
        alias:
          type: string
          minLength: 1
          maxLength: 32
        bio:
          type: string
          nullable: true
          maxLength: 160
        avatar:
          $ref: "#/components/schemas/Avatar"
        createdAt:
          type: string
          format: date-time
        noteCount:
          type: integer
          minimum: 0
        isBlocked:
          type: boolean
          description: True when the requesting persona has blocked or is blocked by this persona.
...
    Report:
      type: object
      required:
        - id
        - targetType
        - targetId
        - category
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
            - post
            - comment
            - persona
        targetId:
          type: string
          format: uuid
        category:
          type: string
          enum:
            - harassment
            - hateSpeech
            - harmfulContent
            - spam
            - sexualContent
            - doxxing
            - impersonation
            - illegalContent
            - other
        details:
          type: string
          nullable: true
          maxLength: 2000
        status:
          type: string
          enum:
            - open
            - resolved
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    ReportCreateRequest:
      type: object
      required:
        - targetType
        - targetId
        - category
      properties:
        targetType:
          type: string
          enum:
            - post
            - comment
            - persona
        targetId:
          type: string
          format: uuid
        category:
          type: string
          enum:
            - harassment
            - hateSpeech
            - harmfulContent
            - spam
            - sexualContent
            - doxxing
            - impersonation
            - illegalContent
            - other
        details:
          type: string
          maxLength: 2000
    Block:
      type: object
      required:
        - id
        - persona
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        persona:
          $ref: "#/components/schemas/Persona"
        createdAt:
          type: string
          format: date-time
    BlockCreateRequest:
      type: object
      required:
        - personaId
      properties:
        personaId:
          type: string
          format: uuid
    ModerationCase:
      type: object
      required:
        - id
        - targetType
        - targetId
        - reportIds
        - status
        - outcome
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
            - post
            - comment
            - persona
        targetId:
          type: string
          format: uuid
        reportIds:
          type: array
          items:
            type: string
            format: uuid
        status:
          type: string
          enum:
            - open
            - underReview
            - resolved
        outcome:
          type: string
          nullable: true
          enum:
            - noAction
            - warn
            - hide
            - remove
            - restore
            - restrictPersona
            - suspendAccount
            - banAccount
          description: Null until the case is resolved.
        notes:
          type: string
          maxLength: 4000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ModerationCaseCreateRequest:
      type: object
      required:
        - targetType
        - targetId
        - reportIds
      properties:
        targetType:
          type: string
          enum:
            - post
            - comment
            - persona
        targetId:
          type: string
          format: uuid
        reportIds:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
    ModerationCaseUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - underReview
            - resolved
        outcome:
          type: string
          enum:
            - noAction
            - warn
            - hide
            - remove
            - restore
            - restrictPersona
            - suspendAccount
            - banAccount
        notes:
          type: string
          maxLength: 4000
...
    CursorPage:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          required:
            - nextCursor
            - hasMore
          properties:
            nextCursor:
              type: string
              nullable: true
            hasMore:
              type: boolean
...
    CursorPageBlock:
      allOf:
        - $ref: "#/components/schemas/CursorPage"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Block"
    CursorPageModerationCase:
      allOf:
        - $ref: "#/components/schemas/CursorPage"
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/ModerationCase"
  responses:
    BadRequest:
      description: The request was malformed or failed basic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the access token is invalid/expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated principal lacks permission for the resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current state (duplicate, idempotency mismatch, etc.).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: The request was syntactically valid but semantically invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Too many requests. Retry after the period indicated by `Retry-After`.
      headers:
        Retry-After:
          description: Seconds until the client may retry.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"